Improves payment method form and validation
Enhances the payment method form by adding labels to the name and description fields for better user experience.

Also, it updates the locale validation rule to ensure that the locale code exists in the languages table, improving data integrity.

Additionally, it excludes 'sandbox' and 'token' from the PayPal gateway initialization to prevent potential configuration issues.

Finally, it handles the case where the payment method config is null when rendering the config form.